header dinamic y documnetation of code
