+++ frontend/src/components/pages/Mitologia_.jsx
@@ -6,6 +6,7 @@ import icon_profile from '../../assets/logo/profile_default.svg';
 import Section from '../history/Section';
 import { useNavigate } from 'react-router-dom';
 
+//Informacion de la seccion mitologia
 const DATA_HISTORY = [
     {
         title: 'Mitología',
@@ -16,25 +17,17 @@ const DATA_HISTORY = [
     }
 ]
 
-
+/**
+ * @description Renderiza las opciones de la seccion mitologia
+ * @returns Mitologia Component
+*/
 const Mitologia_ = () => {
 
     const history =  useNavigate();
-    const [ active_profile, setActiveProfile ] = React.useState(false);
-
-    const handleProfile = () => {
-        setActiveProfile(!active_profile);
-    }
 
     const handleCerrarSesion = (target,pathBack) => {
         history(target, {  state: { pathBack } });
     }
-
-    const handleCerrar = (target) => {
-        document.getElementById('root').classList.remove('remove_gap');
-        history(target);
-    }
-
 
     React.useEffect(() => {
         document.getElementById('root').classList.add('remove_gap');
@@ -42,18 +35,7 @@ const Mitologia_ = () => {
 
     return (
         <>
-            <Header>
-                <div className='_container_profile'>
-                    <a onClick={handleProfile}>
-                        <img src={icon_profile} alt="profile" />
-                        { active_profile &&
-                          <div className='_container_options'>
-                                <a onClick={ () => { handleCerrar('/') }}>Cerrar Sesion</a>
-                          </div>
-                        }
-                    </a>
-                 </div>
-            </Header>
+            <Header login={true} />
             <div className='_container_primary_profile'>
                 <Navegator />
                 <Section title='Mitologia' >
